Fix DNS resolution problems
There are two distinct DNS resolution problems. One, the nodename
reverse DNS resolution. And two, resolution of peers. This attempts
to fix both problems.

Nodename is only required when rabbitmq runs multiple instances on
the same host. Having this set complicates the DNS requirements.
Removing this setting may simplify the DNS problems this charm has
faced.

For peer resolution force the use of hostname rather than dynamically
attempting to resolve one from an IP. Set /etc/hosts with the local
hostname and with each peer.

Standardize the way an IPv4 or IPv6 address is selected throughout
the charm. Standardize the way a hostname is selected throughout the
charm.

Partial-Bug: 1584902
Change-Id: I105eb2684e61a553a52c5a944e8c562945e2a6eb