Wait until clustered before running client hooks
RabbitMQ takes some time to fully cluster. The charm was previously
running amqp-relation-changed hooks whenever they were queued even
if the cluster was not yet complete. This led to split brain
scenarios. Client authentication to one or more nodes could fail.

This change confirms the entire cluster is ready before running
client amqp-relation-changed hooks.

min-cluster-size can now be used to attempt to guarantee the cluster
is ready with the expected number of nodes. If min-cluster-size is
not set the charm will still determine based on the information
available if all the cluster nodes are ready. Single node
deployments are still possible.

Partial-Bug: #1657245
Closes-Bug: #1657176
Change-Id: I870df71869c979e65a3a8764efdf35a746278507