networking: Fix amqp relation network binding
When configuration for 'access-network' is supplied, only the
lead unit in a cluster will present the correct network address
for access to the RMQ cluster; followers always present 'private-address'
as they echo out data from the peer relation or leader storage
on amqp relations, overriding the hostname and private-address
values with the return value of get_host_ip.

Pass 'amqp_relation=True' to get_host_ip to ensure that any supplied
configuration option is correctly used, or that for Juju 2.0, any
network space binding for the amqp relation is used.

This is resolved in master branch by a larger refactoring of the
resolution of addressing for the amqp and cluster relations.

Change-Id: Ia4921becf149243fab40f9f21a26bfbe9d3da332
Closes-Bug: 1605314