Correct port that is exposed for management
For rabbitmq-server version 3+ (Trusty and above) the managment
port is 15672, however the charm exposes the old port 55672. This
change fixes that behaviour. Fwiw the amulet tests are already
expecting the managment interface to be 15672 but tests run with the
OpenStack providor and would have missed that this port was not
being exposed.

Change-Id: Ia61d1c12df9300408d33a1768f546c38b33c5677
Closes-Bug: 1651195