Set clustered atrribute in the cluster's leader
When forming the cluster the leader node (defined by
leader_node_hostname) is the one that won't join a cluster, all the
other peers will join this leader, this prevented the leader from
setting the clustered attribute in the cluster relation.

This patch makes all the units to run the cluster_with() function
and cluster_with() skips 'rabbitmqctl join_cluster' for the nodes
already present in the running section of 'rabbitmqctl cluster_status'
and sets the clustered attribute if it hasn't been set already.

Also cluster_with() is called within the upgrade-charm hook to allow
existing deployments set the attribute if needed.

Change-Id: I234046de279e1c28f7ad640c2cd7c641e864525b
Closes-Bug: 1691510