Refactor amqp_changed to make it less noisy
Currently each and every hook executed results in a sweep of all
amqp relations and their related units to ensure config consistency
both in rabbit and for remote units. This scales very poorly since
rabbtimqctl commands take a while to complete and the majority of
the time they are not needed. This patch aims to reduce the impact
of performing this set of operations and the amount of time it takes
to do them.

Change-Id: Ia060ce34052cd543a63d045944b35e4188279f05
Closes-Bug: 1698340
(cherry picked from commit da3a7063bbcf0c262957149d5c6c2fa3d0525599)