Remove update_clients() call from update-status.
The commit [0] added a explicit call to update_clients
which is executed on every hook run (including update-status),
the update_clients function iterates over all the units
related through the amqp relation and makes a call
to the amqp_changed function.

As the update_status hook interval is set to run every 5 minutes,
isn't convenient to do a call to amqp_changed for every related unit
particularly in large scale deployments, as this might cause
an unwanted workload overhead on the cloud at every tick
of the update-status hook.

Further work needs to be done to improve the overall
handling performed via the update_clients function but
this commit alleviates the problem of leaving the update-status
hook stuck for ~1 hour every time.

[0] 2472e1ca9fbd68cdd50640eadd071177ddc5968e

Change-Id: If9d6ed03d0b8a79fa0210fba57a3d54c2646214d
Closes-Bug: #1717579
Signed-off-by: Jorge Niedbalski <jorge.niedbalski@canonical.com>