Unset ssl_port when disabling ssl
Some Rabbitmq clients gate enabling ssl on the presence of the
ssl_port option in the amqp relation. However, when ssl is disabled
the rabbitmq-server charm does not explicitly unset ssl_port so it
persists down the relation. This change unsets the ssl_port
option along with the other ssl options when they are not needed.

There seemed to be no unit test coverage for ssl_utils so I've
added some as part of this change.

Change-Id: If957899427ec979d09045baca010375ba7273579
Closes-Bug: #1717982
(cherry picked from commit 407cfa80975a170bf2b34b9ea60d0d9e49109971)