Added NRPE monitoring for cluster partition state
The current RabbitMQ charm does not have built-in monitoring for
the RabbitMQ cluster partition state.

This feature was requested by our operations team, but also in
https://bugs.launchpad.net/charm-rabbitmq-server/+bug/1548679

The implementation in this merge request is very similar to the
existing queue length monitoring. The cluster state is retrieved by
running plain rabbitmqctl from cron as root, output is saved into a
data file that is checked by the nrpe plugin script.

Another option could be to use the Management API, but in our
environment it is not enabled.

Closes-Bug: 1548679